+++ src/index.js
@@ -32,6 +32,32 @@ class SSH {
       })
     )
   }
+  async requestShell(): Promise<Object> {
+    const connection = this.connection
+    invariant(connection, 'Not connected to server')
+    return await new Promise(function(resolve, reject) {
+      connection.shell(function(error, shell) {
+        if (error) {
+          reject(error)
+        } else {
+          resolve(shell)
+        }
+      })
+    })
+  }
+  async requestSFTP(): Promise<Object> {
+    const connection = this.connection
+    invariant(connection, 'Not connected to server')
+    return await new Promise(function(resolve, reject) {
+      connection.sftp(function(error, sftp) {
+        if (error) {
+          reject(error)
+        } else {
+          resolve(sftp)
+        }
+      })
+    })
+  }
   async mkdir(path: string): Promise<void> {
     invariant(this.connection, 'Not connected to server')
     await this.exec('mkdir', ['-p', path])
@@ -89,33 +115,7 @@ class SSH {
       })
     })
   }
-  async requestShell(): Promise<Object> {
-    const connection = this.connection
-    invariant(connection, 'Not connected to server')
-    return await new Promise(function(resolve, reject) {
-      connection.shell(function(error, shell) {
-        if (error) {
-          reject(error)
-        } else {
-          resolve(shell)
-        }
-      })
-    })
-  }
-  async requestSFTP(): Promise<Object> {
-    const connection = this.connection
-    invariant(connection, 'Not connected to server')
-    return await new Promise(function(resolve, reject) {
-      connection.sftp(function(error, sftp) {
-        if (error) {
-          reject(error)
-        } else {
-          resolve(sftp)
-        }
-      })
-    })
-  }
-  async getFile(localFile: string, remoteFile: string, givenSftp: ?Object = null) {
+  async getFile(localFile: string, remoteFile: string, givenSftp: ?Object = null): Promise<void> {
     invariant(this.connection, 'Not connected to server')
     invariant(typeof localFile === 'string' && localFile, 'localFile must be a string')
     invariant(typeof remoteFile === 'string' && remoteFile, 'remoteFile must be a string')
@@ -138,12 +138,12 @@ class SSH {
       }
     }
   }
-  async putFile(localFile: string, remoteFile: string, givenSftp: ?Object = null) {
+  async putFile(localFile: string, remoteFile: string, givenSftp: ?Object = null): Promise<void> {
     invariant(this.connection, 'Not connected to server')
     invariant(typeof localFile === 'string' && localFile, 'localFile must be a string')
     invariant(typeof remoteFile === 'string' && remoteFile, 'remoteFile must be a string')
     invariant(!givenSftp || typeof givenSftp === 'object', 'sftp must be an object')
-    invariant(await Helpers.fileExists(localFile), `localFile does not exist at ${localFile}`)
+    invariant(await Helpers.exists(localFile), `localFile does not exist at ${localFile}`)
 
     const that = this
     const sftp = givenSftp || await this.requestSFTP()
@@ -174,7 +174,7 @@ class SSH {
       }
     }
   }
-  async putFiles(files: Array<{ local: string, remote: string }>, givenSftp: ?Object) {
+  async putFiles(files: Array<{ local: string, remote: string }>, givenSftp: ?Object = null): Promise<void> {
     invariant(this.connection, 'Not connected to server')
     invariant(!givenSftp || typeof givenSftp === 'object', 'sftp must be an object')
     invariant(Array.isArray(files), 'files must be an array')
@@ -187,11 +187,9 @@ class SSH {
     }
 
     const sftp = givenSftp || await this.requestSFTP()
-    const promises = []
-    for (let i = 0, length = files.length; i < length; ++i) {
-      const file = files[i]
-      promises.push(this.putFile(file.local, file.remote, sftp))
-    }
+    const promises = files.map(file =>
+      this.putFile(file.local, file.remote, sftp)
+    )
     try {
       await Promise.all(promises)
     } finally {
@@ -199,6 +197,51 @@ class SSH {
         sftp.end()
       }
     }
+  }
+  async putDirectory(localDirectory: string, remoteDirectory: string, givenConfig: Object = {}, givenSftp: ?Object = null): Promise<boolean> {
+    invariant(this.connection, 'Not connected to server')
+    invariant(typeof localDirectory === 'string' && localDirectory, 'localDirectory must be a string')
+    invariant(typeof remoteDirectory === 'string' && remoteDirectory, 'localDirectory must be a string')
+    invariant(await Helpers.exists(localDirectory), `localDirectory does not exist at ${localDirectory}`)
+    invariant((await Helpers.stat(localDirectory)).isDirectory(), `localDirectory is not a directory at ${localDirectory}`)
+    invariant(typeof givenConfig === 'object' && givenConfig, 'config must be an object')
+
+    const sftp = givenSftp || await this.requestSFTP()
+    const config = Helpers.normalizePutDirectoryConfig(givenConfig)
+    const files = (await Helpers.scanDirectory(localDirectory, config.recursive ? 2 : 1, config.validate))
+      .map(function(item) {
+        return Path.relative(localDirectory, item)
+      })
+    const directoriesCreated = new Set()
+
+    const promises = files.map(async file => {
+      const localFile = Path.join(localDirectory, file)
+      const remoteFile = Path.join(remoteDirectory, file)
+      const remoteFileDirectory = Path.dirname(remoteFile)
+      if (!directoriesCreated.has(remoteFileDirectory)) {
+        await this.mkdir(remoteFileDirectory)
+        directoriesCreated.add(remoteFileDirectory)
+      }
+      try {
+        this.putFile(localFile, remoteFile, sftp)
+        config.tick(localFile, remoteFile, null)
+        return true
+      } catch (_) {
+        config.tick(localFile, remoteFile, _)
+        return false
+      }
+    })
+
+    let results
+    try {
+      results = await Promise.all(promises)
+    } finally {
+      if (!givenSftp) {
+        sftp.end()
+      }
+    }
+
+    return results.every(i => i)
   }
   dispose() {
     if (this.connection) {