+++ src/types.js
@@ -16,3 +16,15 @@ export type ConfigGiven = {
   password?: string,
   privateKey?: string,
 }
+
+export type ConfigDirectoryTransfer = {
+  recursive: boolean,
+  tick: ((localPath: string, remotePath: string, error: ?Error) => void),
+  validate: ((localPath: string) => boolean),
+}
+
+export type ConfigDirectoryTransferGiven = {
+  recursive?: boolean,
+  tick?: ((localPath: string, remotePath: string, error: ?Error) => void),
+  validate?: ((localPath: string) => boolean),
+}